requirements.txt
django==3.2.9
djangorestframework==3.12.4
dj-rest-auth==2.1.11
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
django-environ==0.8.1